Added temp loading indicator on sign in
Status bar color changes from white (when on main app background) to black (when in sign up safari view controller created by Firebase) to be visible against its white background.

There’s a bit of a delay, however, and it might be feasiblie to listen for the safariviewcontroller callback in Firebase (somehow) to only trigger the change from white back to black when the controller is dismissed, not when the user is signed in.