Added ErrorHandler, implemented some
Only have localUserCouldNotBeCreatedException, which is not an NSError type, but a normal function.

I’m not currently sure if FIR/GID functions can throw errors, as the ErrorHandler functions are usually called within the *completion block*, when an NSError object has (usually) already been instantiated and subsequently returned.

For now, they’re void.