Start of “two-step” shaing framework
When share button is tapped, the user info table view will enter a selection-based editing mode. The nav bar button items will change to “cancel” and “share”, and allow the user to select multiple properties of the User to share.

For instance, they can share their name and date, which would generate the string:
> “[name] was last in the ilab on [day of week] at [time”!

#dynamism