Parsing Algorithm started (see notes)
Started to develop the parsing algorithm with a corresponding ShareStringParseError ErrorType.

The shareUserInformation function was updated to conform to the new input requirements: namely, that input parameters are now organized by a new enum shareType.

The formal definition is in progress, and only two cases are defined so far (both being erronous).