Progress on Date parsing for user sharing
More defined cases for sharing actions
Conditional casting issues through lines 289-298, partially due to the fact that the algortihm is written in a Playground with Swift 3, not 2.3 (as in this project).

Will be converted when this algortihm is done for the sake of clarity. It’s already confusing enough to write it, nevermind undergo a major language version shift partway through!